Allow 0-stride dimensions for cublas input/output
Currently, if a tensor contains 0-stride dimensions in one of the
last two dimensions, then that tensor will not be used directly as
input/output to cublas and a copy will be made transparently. This
generates copies during matvec operations because matvec clones a
tensor and delegates to matmul, which then copies the tensor because
the clone creates a 0-stride dimension. This change allows 0-stride
dimensions if the dimension size is 1 (and thus no index will stride
in that dimension) and thus eliminates allocations/frees and copies
when calling matvec.