Optimize resample_poly with a multi-kernel approach
Add additional resample_poly kernels that are more optimal in
specific cases (e.g., in the case of a pure downsampler). Currently,
there are three kernels, but only two of them are in use. There are
some cases for which the third is fastest, but a pattern for when
that will be the case is not yet clear.