Zero-initialize batches field in CUB params (#527)
For some CUB reduction types, the batches field in the CUB params struct
is uninitialized. The field is not needed for the reduction, but it is
included when performing cache lookups, so the uninitialized value could
cause false negatives (cache misses) and additional plan creations. This
did not cause correctness issues, but could impact performance. Thus, zero
initialize the batches field to prevent this. The CUB cache is currently
disabled by default, but can be enabled at build time by the user.