Prevent potential oob read in matxOpTDKernel (#586)
During the last iteration when expanding the linear index
into the N-D tensor index, there may be an out-of-bounds
read access into the sizes array. The result of the read
is unused and the iteration count is known at compile time,
so the read may be removed by the compiler. Extract the
last iteration into an epilogue to guarantee no out-of-bounds
access.

Signed-off-by: Thomas Benson <tbenson@nvidia.com>