Add workaround for pre-11.8 CTK smem init errors (#673)
Static shared memory cannot be initialized by constructors, so the lack of
trivial constructors in the std::complex types causes older CUDA toolkit
releases to report an error on static shared memory declarations of
complex types. Treat the memory as aligned byte array and cast to the
appropriate type as a workaround.

Fixes issue #669.