Use cublas<t>matinvBatched() for N <= 32
Use the cublas<t>matinvBatched() family of functions to invert linear systems
of size N <= 32. This has two advantages over the more general pair of
getrfBatched() and getriBatched() functions:

1. Higher performance with the single kernel than with split kernels.
2. The matinv functions support in-place transforms and do not modify the
input in the case of out-of-place transforms, so we do not need a temporary
input work buffer if the input is a tensor view.