Create temp contiguous tensors if needed for sort (#757)
The sort implementation currently requires contiguous tensor view
inputs. The run-time check for that requirement is currently disabled
for release builds and thus those builds can produce incorrect results
for non-contiguous tensors.

This change creates temporary contiguous allocations and copies the
input when necessary to support sort(). This has the side effect of
adding some operator input support for sort() because the operator
will execute to populate the temporary tensor.